speed limit modified, not for tanks for some reason

+++ src/main/java/mcheli/tank/MCH_EntityTank.java
@@ -930,7 +930,7 @@ public class MCH_EntityTank extends MCH_EntityAircraft {
       if(motion1 > (double)speedLimit) {
          super.motionX *= (double)speedLimit / motion1;
          super.motionZ *= (double)speedLimit / motion1;
-         motion1 = (double)speedLimit;
+         motion1 = speedLimit;
       }
 
       if(motion1 > prevMotion && super.currentSpeed < (double)speedLimit) {
